requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
pyyaml==6.0.1
python-dotenv==1.0.1
weaviate-client==4.8.1
sentence-transformers==3.0.1
transformers==4.44.2
redis==5.0.8
celery==5.4.0
requests==2.32.3
mlflow==2.15.1

